+++ src/sexp/environment.rs
@@ -4,6 +4,8 @@ use savvy_ffi::{R_NilValue, R_UnboundValue, Rboolean_FALSE, Rboolean_TRUE, SEXP}
 
 use crate::Sexp;
 
+use super::utils::str_to_symsxp;
+
 /// An environment.
 pub struct EnvironmentSexp(pub SEXP);
 
@@ -14,7 +16,7 @@ impl EnvironmentSexp {
     }
 
     pub fn get<T: AsRef<str>>(&self, name: T) -> crate::error::Result<Option<crate::Sexp>> {
-        let sym = str_to_symsexp(name)?.ok_or("name must not be empty")?;
+        let sym = str_to_symsxp(name)?.ok_or("name must not be empty")?;
 
         // Note: since this SEXP already belongs to an environment, this doesn't
         // need protection.
@@ -30,7 +32,7 @@ impl EnvironmentSexp {
     }
 
     pub fn contains<T: AsRef<str>>(&self, name: T) -> crate::error::Result<bool> {
-        let sym = str_to_symsexp(name)?.ok_or("name must not be empty")?;
+        let sym = str_to_symsxp(name)?.ok_or("name must not be empty")?;
 
         let res = unsafe {
             crate::unwind_protect(|| savvy_ffi::Rf_findVarInFrame3(self.0, sym, Rboolean_FALSE))?
@@ -55,22 +57,6 @@ impl EnvironmentSexp {
 
         Ok(())
     }
-}
-
-// Note: the result is not protected (although symbol is probably not GC-ed?)
-pub(crate) fn str_to_symsexp<T: AsRef<str>>(name: T) -> crate::error::Result<Option<SEXP>> {
-    let name = name.as_ref();
-    if name.is_empty() {
-        return Ok(None);
-    }
-
-    let name_cstr = match CString::new(name) {
-        Ok(cstr) => cstr,
-        Err(e) => return Err(crate::error::Error::new(&e.to_string())),
-    };
-    let sym = unsafe { crate::unwind_protect(|| savvy_ffi::Rf_install(name_cstr.as_ptr())) }?;
-
-    Ok(Some(sym))
 }
 
 // conversions from/to EnvironmentSexp ***************
