+++ src/sexp/utils.rs
@@ -1,4 +1,7 @@
-use std::{ffi::CStr, os::raw::c_char};
+use std::{
+    ffi::{CStr, CString},
+    os::raw::c_char,
+};
 
 use savvy_ffi::{cetype_t_CE_UTF8, Rf_mkCharLenCE, Rf_xlength, R_CHAR, SEXP};
 
@@ -47,4 +50,20 @@ pub(crate) unsafe fn charsxp_to_str(v: SEXP) -> &'static str {
             .to_str()
             .unwrap_or_default()
     }
+}
+
+// Note: the result is not protected (although symbol is probably not GC-ed?)
+pub(crate) fn str_to_symsxp<T: AsRef<str>>(name: T) -> crate::error::Result<Option<SEXP>> {
+    let name = name.as_ref();
+    if name.is_empty() {
+        return Ok(None);
+    }
+
+    let name_cstr = match CString::new(name) {
+        Ok(cstr) => cstr,
+        Err(e) => return Err(crate::error::Error::new(&e.to_string())),
+    };
+    let sym = unsafe { crate::unwind_protect(|| savvy_ffi::Rf_install(name_cstr.as_ptr())) }?;
+
+    Ok(Some(sym))
 }
